Add a regression test for nullable range indexer (#33803)
Closes #31748
